FIX BUILD 8
Nerfed the gravity obstacle

Depending the gravity of the player, the gravity portal will change it´s color

Added the Gravity_Obstacle Script, it controls the Gravity Manager and call the change gravity method when collides with the player
