+++ Assets/Prefabs/Obstacles/Obstacle_Container8.prefab
@@ -83,7 +83,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4314861392884902993}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 5.8133, z: 0}
+  m_LocalPosition: {x: 0, y: 7, z: 0}
   m_LocalScale: {x: 1, y: 6.6265, z: 1}
   m_Children: []
   m_Father: {fileID: 4314861393131942928}
@@ -192,7 +192,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4314861392925287600}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: -5.9119, z: 0}
+  m_LocalPosition: {x: 0, y: -7, z: 0}
   m_LocalScale: {x: 1, y: 6.824, z: 1}
   m_Children: []
   m_Father: {fileID: 4314861393131942928}
@@ -355,8 +355,8 @@ GameObject:
   - component: {fileID: 8936586823034850087}
   - component: {fileID: 8936586823034850085}
   - component: {fileID: 8936586823034850086}
-  - component: {fileID: 8936586823034850084}
   - component: {fileID: 8936586823034850091}
+  - component: {fileID: 559988852206653255}
   m_Layer: 0
   m_Name: Capsule
   m_TagString: Untagged
@@ -445,19 +445,6 @@ CapsuleCollider2D:
   m_Offset: {x: 0, y: 0}
   m_Size: {x: 1, y: 2}
   m_Direction: 0
---- !u!114 &8936586823034850084
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8936586823034850080}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d776e1dce3f324c4ba35b61d5f878164, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  gravityAxis: 1
 --- !u!50 &8936586823034850091
 Rigidbody2D:
   serializedVersion: 4
@@ -479,3 +466,15 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
+--- !u!114 &559988852206653255
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8936586823034850080}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 42966ec0e90669748960d2af556ec970, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
