+++ Assets/Scripts/Enviroment/Obstacle/Gravity_Obstacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 42966ec0e90669748960d2af556ec970
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
